Some Quality of Life Improvements (#254)
This adds a two methods to Result, both of which are taken from Rust:
`is_ok()` and `map_or_else(Function<T, U> resultFn, Function<E, U>
errorFn)`

This also adds a `result-test-utils` package and sets it up for
publishing. This includes an implementation of `AbstractAssert` for
Result so that one can use a custom `Result` `assertThat()` with some
nice helpful assertion methods.